+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for pygarn
 
+
+## 0.2
+- added UnfoldSubGraph operation
+- added AddCompleteGraph operation with its backward variant
+- added new optional parameter for backward "return_fuzzy" to return potential original graphs instead of a trial one
+
 ## 0.1
 - first exemplary operations with common interface to add or duplicate a graph
 - selectors for various purposes based on random uniform selection or based on degree of vertices